+++ src/pykeen/nn/utils.py
@@ -14,7 +14,8 @@ import requests
 import torch
 
 from ..constants import PYKEEN_MODULE
-from ..utils import nested_get
+from ..typing import OneOrSequence
+from ..utils import nested_get, upgrade_to_sequence
 from ..version import get_version
 
 __all__ = [
@@ -22,6 +23,7 @@ __all__ = [
     "adjacency_tensor_to_stacked_matrix",
     "use_horizontal_stacking",
     "WikidataCache",
+    "ShapeError",
 ]
 
 logger = logging.getLogger(__name__)
@@ -324,3 +326,42 @@ class WikidataCache:
             the description for each Wikidata entity
         """
         return self._get(ids=ids, component="description")
+
+
+class ShapeError(ValueError):
+    """An error for a mismatch in shapes."""
+
+    def __init__(self, shape: OneOrSequence[int], reference: OneOrSequence[int]) -> None:
+        """
+        Initialize the error.
+
+        :param shape: the mismatching shape
+        :param reference: the expected shape
+        """
+        super().__init__(f"shape {shape} does not match expected shape {reference}")
+
+    @classmethod
+    def verify(cls, shape: OneOrSequence[int], reference: Optional[OneOrSequence[int]]) -> Sequence[int]:
+        """
+        Raise an exception if the shape does not match the reference.
+
+        This method normalizes the shapes first.
+
+        :param shape:
+            the shape to check
+        :param reference:
+            the reference shape. If None, the shape always matches.
+
+        :raises ShapeError:
+            if the two shapes do not match.
+
+        :return:
+            the normalized shape
+        """
+        if reference is None:
+            return
+        reference = upgrade_to_sequence(reference)
+        shape = upgrade_to_sequence(shape)
+        if reference != shape:
+            raise cls(shape=shape, reference=reference)
+        return shape
